project/python/requirements.txt
boto3
Pillow
pillow-heif
requests